ai_resume_analyzer/requirements.txt
openai
fastapi
uvicorn
python-dotenv
PyPDF2
python-multipart
requests